requirements/base.txt
asgi-correlation-id==4.3.4
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.0
dependency-injector==4.46.0
pydantic_settings==2.9.1
uuid7==0.1.0
motor==3.7.1